Unit tests (#61)
* device orientation tests

* dom debugger tests

* dom snapshot tests

* dom storage tests

+++ cdtp/device_orientation/command.go
@@ -18,13 +18,13 @@ https://chromedevtools.github.io/devtools-protocol/tot/DeviceOrientation/#method
 */
 type SetOverrideParams struct {
 	// Mock alpha.
-	Alpha float64 `json:"alpha"`
+	Alpha int64 `json:"alpha"`
 
 	// Mock beta.
-	Beta float64 `json:"beta"`
+	Beta int64 `json:"beta"`
 
 	// Mock gamma.
-	Gamma float64 `json:"gamma"`
+	Gamma int64 `json:"gamma"`
 }
 
 /*
